Unificar clases en un archivo
